dictinary/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.4
django-ckeditor==6.7.0
django-cors-headers==4.2.0
django-filter==23.2
django-jazzmin==2.6.0
django-js-asset==2.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.8.1
environs==9.5.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
marshmallow==3.20.1
packaging==23.1
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
whitenoise==6.5.0